fix: 나의 채팅 목록 필터 적용
