+++ src/components/channel/ChannelSelector.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex, HStack, useRadioGroup } from '@chakra-ui/react';
+import { Box, Grid, HStack, useRadioGroup } from '@chakra-ui/react';
 import { CATEGORIES, CATEGORY_COLOR_SCHEMES } from '../../constants/channel';
 import ChannelRadio from './ChannelRadio';
 import ScrollContainer from 'react-indiana-drag-scroll';
@@ -14,42 +14,40 @@ const ChannelSelector = () => {
   const group = getRootProps();
 
   return (
-    <>
-      <Flex>
-        <Box
-          display="flex"
-          justifyContent="center"
-          alignContent="center"
-          bg="#F4F4F4"
-          w="full"
-          mt="4"
-          mb="4"
-          color="black"
-          border="1px solid #cbcbcb"
-          borderRadius="4"
-          boxShadow="md"
-          p="4"
-          overflow="auto"
-        >
-          <ScrollContainer className="scroll-container">
-            <HStack {...group} align="center" spacing={4} mb="1">
-              {options.map((option) => {
-                const radio = getRadioProps({ value: option.value });
-                return (
-                  <ChannelRadio
-                    key={option.name}
-                    {...radio}
-                    customColor={CATEGORY_COLOR_SCHEMES[option.name]}
-                  >
-                    {option.name}
-                  </ChannelRadio>
-                );
-              })}
-            </HStack>
-          </ScrollContainer>
-        </Box>
-      </Flex>
-    </>
+    <Grid>
+      <Box
+        display="flex"
+        justifyContent="center"
+        alignContent="center"
+        bg="#F4F4F4"
+        w="full"
+        mt="4"
+        mb="4"
+        color="black"
+        border="1px solid #cbcbcb"
+        borderRadius="4"
+        boxShadow="md"
+        p="4"
+        overflow="auto"
+      >
+        <ScrollContainer className="scroll-container">
+          <HStack {...group} align="center" spacing={4} mb="1">
+            {options.map((option) => {
+              const radio = getRadioProps({ value: option.value });
+              return (
+                <ChannelRadio
+                  key={option.name}
+                  {...radio}
+                  customColor={CATEGORY_COLOR_SCHEMES[option.name]}
+                >
+                  {option.name}
+                </ChannelRadio>
+              );
+            })}
+          </HStack>
+        </ScrollContainer>
+      </Box>
+    </Grid>
   );
 };
