+++ src/components/channel/MyChannels.tsx
@@ -2,11 +2,13 @@ import { Grid, GridItem, HStack } from '@chakra-ui/react';
 import { useMyChannels } from '../../hooks/useMyChannels';
 import ChannelCard from './ChannelCard';
 import LoadingSkeleton, { skeletons } from './LoadingSkeleton';
+import useFilteredChannels from '../../hooks/useFilteredChannels';
 
 const MyChannels = () => {
-  const { data: myChannels, isLoading } = useMyChannels();
+  const { data: channels, isLoading } = useMyChannels();
+  const filteredChannels = channels ? useFilteredChannels(channels) : [];
 
-  if (isLoading || !myChannels)
+  if (isLoading || !channels)
     return (
       <HStack>
         {skeletons.map((_i, index) => (
@@ -17,10 +19,10 @@ const MyChannels = () => {
 
   return (
     <Grid gap={4} gridTemplateColumns={'repeat(auto-fit, minmax(220px, 1fr))'}>
-      {myChannels.length === 0 ? (
+      {filteredChannels.length === 0 ? (
         <div>채팅방이 없습니다.</div>
       ) : (
-        myChannels.map((channel) => (
+        filteredChannels.map((channel) => (
           <GridItem key={channel.id}>
             <ChannelCard channel={channel} />
           </GridItem>
